move into `src/server/main.rs` to keep the module clean
